remove row attribute prop
