+++ app/src/lib/design-system/components/RQNetworkTable/index.tsx
@@ -1,11 +1,11 @@
-import React, { useMemo, useEffect, useState } from "react";
+import React, { useMemo, useEffect, useState, useRef } from "react";
 import { GenericNetworkTable, GenericNetworkTableProps } from "./GenericNetworkTable";
 import { RQSessionAttributes } from "@requestly/web-sdk";
 import { secToMinutesAndSeconds } from "utils/DateTimeUtils";
 import { getOffset } from "views/features/sessions/SessionViewer/NetworkLogs/helpers";
 import { RQNetworkLog } from "./types";
 import { AiFillCaretRight } from "@react-icons/all-files/ai/AiFillCaretRight";
-import usePointerAutoScroll from "./hooks/usePointerAutoScroll";
+import useFocusedAutoScroll from "./useFocusedAutoScroll";
 import "./RQNetworkTable.css";
 
 export interface RQNetworkTableProps {
@@ -28,7 +28,8 @@ export const RQNetworkTable: React.FC<RQNetworkTableProps> = ({
   autoScroll = false,
 }) => {
   const [activeLogId, setActiveLogId] = useState(null);
-  const [containerRef, onScroll] = usePointerAutoScroll(activeLogId);
+  const containerRef = useRef(null);
+  const onScroll = useFocusedAutoScroll(containerRef, activeLogId);
 
   const extraColumns: GenericNetworkTableProps<RQNetworkLog>["extraColumns"] = useMemo(
     () => [
@@ -89,7 +90,6 @@ export const RQNetworkTable: React.FC<RQNetworkTableProps> = ({
         autoScroll={autoScroll}
         tableRef={containerRef}
         onTableScroll={onScroll}
-        rowAttributes={(log: RQNetworkLog) => ({ "data-log-id": log.id })}
       />
     </div>
   );
